Removed crop table and changed crop_id to crop_name
