+++ app/model/farm.py
@@ -1,28 +1,32 @@
 import time
-from model.util import id
+from .util import id
 from google.appengine.ext import ndb
 
 
-class Price(ndb.Model):
+class Farm(ndb.Model):
+    ACTION = ['harvest', 'plant', 'sell']
+
     # relationship
-    crop_id = ndb.StringProperty(required=True)
     district_id = ndb.StringProperty(required=True)
 
-    # price properties
-    price = ndb.IntegerProperty(required=True)
+    # farm properties
+    action = ndb.StringProperty(required=True, choices=set(ACTION))
+    crop_name = ndb.StringProperty(required=True)
+    quantity = ndb.IntegerProperty(required=True)
     ts_created = ndb.DateTimeProperty(auto_now_add=True)
     ts_updated = ndb.DateTimeProperty(auto_now=True)
 
     def toJson(self):
         return {
             'id': self.key.id(),
-            'crop_id': self.crop_id,
             'district_id': self.district_id,
-            'price': self.price,
+            'action': self.action,
+            'crop_name': self.crop_name,
+            'quantity': self.quantity,
             'ts_created': int(time.mktime(self.ts_created.timetuple()) * 1000),
             'ts_updated': int(time.mktime(self.ts_updated.timetuple()) * 1000),
         }
 
     @staticmethod
     def id():
-        return 'P{}'.format(id.generate())
+        return 'F{}'.format(id.generate())
