--- app/model/farm/crop.py
@@ -1,22 +1,0 @@
-import time
-from model.util import id
-from google.appengine.ext import ndb
-
-
-class Crop(ndb.Model):
-
-    name = ndb.StringProperty(required=True)
-    ts_created = ndb.DateTimeProperty(auto_now_add=True)
-    ts_updated = ndb.DateTimeProperty(auto_now=True)
-
-    def toJson(self):
-        return {
-            'id': self.key.id(),
-            'name': self.name,
-            'ts_created': int(time.mktime(self.ts_created.timetuple()) * 1000),
-            'ts_updated': int(time.mktime(self.ts_updated.timetuple()) * 1000),
-        }
-
-    @staticmethod
-    def id():
-        return 'C{}'.format(id.generate())
